requirements.txt
gym
numpy
tensorboard
torchvision
opencv-python
matplotlib